Issue #2 - Tests run both locally and in Docker
