+++ web/datacatalog/app.py
@@ -1,7 +1,8 @@
 from aiohttp import web
 
-from . import systemhealth, index, action_api, config
-
+from datacatalog import systemhealth, index, action_api, config
+from datacatalog.search import Search
+from datacatalog.datastore import DataStore
 
 def get_app():
     """
@@ -11,9 +12,13 @@ def get_app():
 
     app.router.add_get('/', index.handle)
     app.router.add_get('/system/health', systemhealth.handle)
-
     app.router.add_get('/datacatalog/api/3/action/{action}', action_api.handle)
 
-    app['config'] = config.load()
+    catalog_config = config.load()
+    app['config'] = catalog_config
+
+    # Initialize plugins
+    app['datastore'] = DataStore(catalog_config).implementation
+    app['search'] = Search(catalog_config).implementation
 
     return app
