+++ web/tests/requirements.txt
@@ -1,1 +1,6 @@
-pytest-aiohttp
+attrs==17.3.0
+pluggy==0.6.0
+py==1.5.2
+pytest==3.3.1
+pytest-aiohttp==0.3.0
+six==1.11.0
